+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change log
 
+## Version 0.1.11
+
+## Added
+
+None
+
+## Changed
+
+#### create function does not require log name any more.
+
+`gracelog.create([*logName])` does not require a `string` `logName`.
+
+If `logName` is not provided, `gracelog` will default the log name to the file name of the script calling `.create()`.
+
 ## Version 0.1.10
 
 ## Added